# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RateMaster API",
    "description": "Revenue/pricing recommendation SaaS for hotels",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/market/snapshot": {
      "post": {
        "tags": [
          "market"
        ],
        "summary": "Create Market Snapshot",
        "description": "Store market snapshot (manual entry or from adapter).",
        "operationId": "create_market_snapshot_api_v1_market_snapshot_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_market_snapshot_api_v1_market_snapshot_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/market/snapshots": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "List Market Snapshots",
        "description": "List market snapshots for property.",
        "operationId": "list_market_snapshots_api_v1_market_snapshots_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Property Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/import-csv": {
      "post": {
        "tags": [
          "market"
        ],
        "summary": "Import Market Csv",
        "description": "Import market/rate-shop CSV. Columns: stay_date or date, compset_avg or adr.",
        "operationId": "import_market_csv_api_v1_market_import_csv_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_market_csv_api_v1_market_import_csv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/model-registry/models": {
      "get": {
        "tags": [
          "model-registry"
        ],
        "summary": "List Models",
        "description": "List registered models - optionally filtered by property.",
        "operationId": "list_models_api_v1_model_registry_models_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "property_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Property Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/model-registry/models/{model_id}/activate": {
      "post": {
        "tags": [
          "model-registry"
        ],
        "summary": "Activate Model",
        "description": "Activate a model version (rollback). Deactivates other versions with same model_name and property.",
        "operationId": "activate_model_api_v1_model_registry_models__model_id__activate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_create_market_snapshot_api_v1_market_snapshot_post": {
        "properties": {
          "property_id": {
            "type": "string",
            "title": "Property Id"
          },
          "compset_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compset Avg"
          },
          "compset_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compset Min"
          },
          "compset_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compset Max"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "manual"
          },
          "stay_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stay Date"
          }
        },
        "type": "object",
        "required": [
          "property_id"
        ],
        "title": "Body_create_market_snapshot_api_v1_market_snapshot_post"
      },
...
      "Body_import_market_csv_api_v1_market_import_csv_post": {
        "properties": {
          "property_id": {
            "type": "string",
            "title": "Property Id"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "property_id",
          "file"
        ],
        "title": "Body_import_market_csv_api_v1_market_import_csv_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}